Loaders re-org
