+++ json-schema-generator/src/main/java/io/micronaut/jsonschema/generator/loaders/FileProcessor.java
@@ -15,27 +15,19 @@
  */
 package io.micronaut.jsonschema.generator.loaders;
 
+import io.micronaut.jsonschema.generator.SourceGenerator;
 import io.micronaut.jsonschema.generator.utils.SourceGeneratorConfig;
 import io.micronaut.jsonschema.model.Schema;
 
 import java.io.File;
 import java.io.IOException;
-import java.io.InputStream;
-import java.net.URI;
-import java.net.http.HttpClient;
-import java.net.http.HttpRequest;
-import java.net.http.HttpResponse;
 import java.nio.file.Path;
-import java.util.List;
-import java.util.regex.Pattern;
-import java.util.regex.Matcher;
+import java.util.Optional;
+
+import static io.micronaut.core.util.StringUtils.capitalize;
+import static io.micronaut.jsonschema.generator.aggregator.TypeAggregator.getCamelCaseName;
 
 public class FileProcessor {
-
-    // configurable by user
-    private static List<String> allowedUrlPatterns = List.of(
-        "^https://.*/.*.json"
-    );
 
     /**
      * Loads a JSON schema based on the configuration.
@@ -48,7 +40,7 @@ public class FileProcessor {
      */
     public static Schema getJsonSchema(SourceGeneratorConfig config) {
         try {
-            SchemaLoader loader = null;
+            SchemaLoader loader;
             if (config.inputStream() != null) {
                 loader = new StreamLoader(config.inputStream());
             } else if (config.jsonUrl() != null && !config.jsonUrl().isBlank()) {
@@ -62,60 +54,6 @@ public class FileProcessor {
         } catch (RuntimeException e) {
             throw new RuntimeException("Error loading JSON schema", e);
         }
-    }
-
-    /**
-     * Downloads a file from the given URL as an InputStream, verifying the URL against
-     * a list of allowed patterns before attempting the download. The URL must end with ".json".
-     *
-     * @param fileUrl The URL to download the file from.
-     * @return An InputStream containing the file content.
-     * @throws IOException If the URL does not match any of the allowed patterns or does not end with ".json",
-     *                   or if an error occurs during download.
-     */
-    public static InputStream downloadAsStream(String fileUrl) throws IOException, InterruptedException {
-        // Verify that the URL matches at least one of the allowed patterns and ends with ".json"
-        if (!isValidUrl(fileUrl)) {
-            throw new IllegalArgumentException("URL does not match any of the allowed patterns or does not starts with https:// and end with .json.");
-        }
-
-        // Create HTTP client and request
-        HttpClient client = HttpClient.newHttpClient();
-        HttpRequest request = HttpRequest.newBuilder()
-            .uri(URI.create(fileUrl))
-            .build();
-
-        // Send the request and return the body as an InputStream
-        HttpResponse<InputStream> response = client.send(request, HttpResponse.BodyHandlers.ofInputStream());
-        return response.body();
-    }
-
-    /**
-     * Validates if the provided URL matches any of the allowed patterns and ends with ".json".
-     *
-     * @param fileUrl The URL to validate.
-     * @return true if the URL matches at least one allowed pattern and ends with ".json", false otherwise.
-     */
-    public static boolean isValidUrl(String fileUrl) {
-        // Check if the URL ends with .json
-        if (!fileUrl.endsWith(".json")) {
-            return false;
-        }
-
-        // Check if the URL starts with http safe
-        if (!fileUrl.startsWith("https://")) {
-            return false;
-        }
-
-        // Check if the URL matches any of the allowed patterns
-        for (String pattern : allowedUrlPatterns) {
-            Pattern compiledPattern = Pattern.compile(pattern);
-            Matcher matcher = compiledPattern.matcher(fileUrl);
-            if (matcher.matches()) {
-                return true;
-            }
-        }
-        return false;
     }
 
     /**
@@ -148,12 +86,22 @@ public class FileProcessor {
         return outputFile;
     }
 
-    public static List<String> getAllowedUrlPatterns() {
-        return allowedUrlPatterns;
-    }
+    public static String getFileName(Schema schema, Optional<String> topLevelName) {
+        String fileName;
+        if (topLevelName.isPresent() && !topLevelName.get().isEmpty()) {
+            fileName = topLevelName.get();
+        } else if (schema.hasTitle()) {
+            fileName = capitalize(getCamelCaseName(schema.getTitle()));
+        } else {
+            fileName = "SchemaFile"; // default
+        }
 
-    public static void setAllowedUrlPatterns(List<String> allowedUrlPatterns) {
-        FileProcessor.allowedUrlPatterns = allowedUrlPatterns;
+        switch (SourceGenerator.getLanguage()) {
+            case KOTLIN: fileName += ".kt"; break;
+            case GROOVY: fileName += ".groovy"; break;
+            default: fileName += ".java"; break;
+        }
+        return fileName;
     }
 }
